Add a position-aware wrapper around AnsiTerminalWriter
When generating output targeted for a specific terminal width, it is
important to know the current position in order to appropriately shorten
the message still to be added to the current line. So make it possible to
add this functionality to the terminal writer itself, to avoid too many
lengthy position computations at call site.

--
Change-Id: I03400b9544c32567fc6ea7ab35e742c4ccd7b610
Reviewed-on: https://bazel-review.googlesource.com/#/c/3373
MOS_MIGRATED_REVID=119946982