Delete the defaultMultipleValue field from options and refactor the logic for retrieving the default values of options.
The field defaultMultipleValue was introduced in commit 4d84ec9a48803c1c097396fa41be1d6e8de9a5ad to allow defining a
default value for options that set allowMultiple. However due to the limitations of
the optionsParser end up being not useful since we cannot guarantee that an option
that allows multiple has a converter that returns a list of values.
Thus this CL deletes code that may confuse even more and clarifies the mechanism
that the options currently use to obtain their default values.

--
MOS_MIGRATED_REVID=120317261