Fix a race condition introduced in unknown commit (that is, the January of 2009!).
If a "blaze clean --expunge" was run concurrently with another command (that was waiting for the lock), it's possible that the clean command deletes the lock file, the new server starts up, then the JVM shutdown hooks delete the PID files from the *new* server.

There is still a slight possibility of a race condition if the lock is deleted then IOException occurs which prevents the BlazeShutdownException from being raised, but I'd rather not introduce another channel from command implementations to RPCServer to close that loophole.

This issue was triggered by commit f03fe8db97af8f8b1f9fb4121fdfef2869e9eeab, after which the PID files for the new server are written a bit more early, thus increasing the time window in which the race condition can happen.

--
MOS_MIGRATED_REVID=120805163