Avoid nested BatchStreamedCallbacks in SkyQueryEnvironment
Previously each call to SkyQueryEnvironment.eval wrapped the provided
callback in a new batched callback. Nesting batched callbacks results
in unnecessary work.

With this change, SkyQueryEnvironment will construct one batched
callback in evaluateQuery (the top-level function called during query
evaluation). Because batched callbacks need to be flushed using
processLastPending(), it was convenient not to call through to the
base class's implementation of evaluateQuery, but to inline it in
SkyQueryEnvironment and add the processLastPending call at the
appropriate place.

This also clears the way for future customizations to
SkyQueryEnvironment's implementation of evaluateQuery.

--
MOS_MIGRATED_REVID=125477770