experimental UI: consider progress bar refreshed on every redraw
Whenever the UI has to write a line to the progress buffer, it removes
the progress bar, write the line and re-adds the progress bar. Of course,
the re-added progress bar is the one that represents the up-to-date state
at the time of it being re-added. Therefore, this also is a refresh of the
progress bar. So update the time stamp tracking the last refresh. In this
way rate-limits on refreshes are honored more correctly; also, requirements
of the progress bar having to be refreshed after a certain point in time
that have been fulfilled by that implicit refresh are recognized as such,
keeping the output less flickering.

--
Change-Id: I44bcd85a149cb88fcde126119100a6e1ff1ba45d
Reviewed-on: https://bazel-review.googlesource.com/#/c/3873
MOS_MIGRATED_REVID=125664608