Add a javac_supports_workers attribute to the java_toolchain rule.
If this is set to true (which it is by default), JavaCompilationHelper will set the "supports-workers" tag on the execution info for the Javac SpawnAction. If it is explicitly set to false, this tag will be absent.

In a follow-up CL, the WorkerSpawnStrategy will check whether this tag is present and fallback to non-worker execution if not.

This is needed to safely enable workers by default in Bazel without breaking the older JDK7 JavaBuilder, which does not support workers yet.

--
MOS_MIGRATED_REVID=126290991