experimental UI: enforce a minimal progress rate limiting if curses unavailable
If the progress bar cannot be updated in place, every additional writing
of the progress bar clutters up the log. Of course, this could be controlled
by adding --show_progress_rate_limit or even --noshow_progress options to the
invocation; however, it turns out that users expect a bazel to just automatically
reduce its output rate. Therefore, specify a minimal value for the progress rate
limit of 2 seconds if the progress bar cannot be updated in place. Also, increase
the time rate for purely time-based updates to 20s in this case.

Moreover, only automatically redraw the progress bar after an INFO/WARNING/... message,
if it can be updated in place.

--
Change-Id: I7d06aa04acec457ec0a5d20ac32ede4cf80f7170
Reviewed-on: https://bazel-review.googlesource.com/#/c/3944
MOS_MIGRATED_REVID=126387987