ExperimentalEventHandler: remove deadlock in update thread
When the progress bar has to be updated, but we cannot do so immediately, e.g.,
due to rate limiting, we start a thread to do this eventually (the "update
thread"). That thread gets stopped if either we realize during an update of the
progress bar that it is no longer needed, or if we get informed about the
completion of the build. Now, since the update thread does updates of the
progress bar (that is its whole purpose) it can get the duty to stop itself.
To avoid the deadlock when the thread there blockingly waits for itself
to terminate, just skip that termination and let the update thread run
till the end of the build.

--
Change-Id: Iba08806f5e416f1fa1825969eceaf6cf54c24e6c
Reviewed-on: https://bazel-review.googlesource.com/#/c/4010
MOS_MIGRATED_REVID=126906395