Prepare JavaBuilder to accept an Aspect's name, in addition to a label, of the owner of a jar.
The motivation is strict-deps checking for Aspected rules such as java_proto_library. With this additional information, JavaBuilder will be able to better tell the user which rule is missing in the 'deps' attribute of the target being built.

Note that the AspectParameters aren't passed right now, because (1) serializing them adds complexity and (2) they're unneeded for JavaBuilder right now.

--
MOS_MIGRATED_REVID=127993845