Move designs/ directory out of docs/ and into site root.
For versioning Bazel's documentation, we only want to version the pages under
the Documentation portion of the site. Since Design Docs are more meant for
Bazel developers and are generally not continually updated for each release,
we should not version the design docs. Moving the directory for the design
docs out of docs/ will also simplify the change for versioning Bazel's docs.

--
MOS_MIGRATED_REVID=128788588