Rollback of commit b5c5158c98ff75592e4ef3c399bff6aa97c71e5c.
*** Reason for rollback ***

This change causes android_integration_test failing on Mac.

*** Original change description ***


Sets SONAME on shared objects in Android binaries.

Adds a linker flag to set the internal DT_SONAME. This fixes #1578 for
SDK 24 and removes the warnings for previous SDKs. There is no need to
set the linker flag for android_librarys that depend on native code,
because the linker flag will be set by the android_binarys that depend
on that android_library.

--
MOS_MIGRATED_REVID=129074607