Dynamic configs: also include fragments needed to resolve the rule's late-bound attributes
To be clear: late-bound attributes introduce two new classes of fragments. First is
those needed by Attribute.LateBoundDefault to figure out what the attribute's value is.
Second is whatever's needed by the actual dep (for label-typed attributes) and its deps.

This change only adds the first. The second is harder since we can't determine
it until we're already in the analysis phase.

--
MOS_MIGRATED_REVID=129117755