Rollback of commit 43ad51f5ff4bc0de4ee800ddfbbd25117f2a5195.
*** Reason for rollback ***

I want to do this in a way that exposes the data attribute only for android_test and android_binary, so I'll add it to the deploy manifest instead.

*** Original change description ***

Expose test "data" attribute to IDE.

This is needed to deploy android_tests to the device for testing. For
symmetry we expose it for all tests rules.

--
MOS_MIGRATED_REVID=129926351