Gracefully handle non existing test descriptions
This change is necessary due to recent move away from Guava immutable collections (which returns null in case of non-existing item) to Java native ones (some of which throws exceptions).

One such use case where it was breaking the code was tests that uses junitparam, which might be an indication of another problem.

--
MOS_MIGRATED_REVID=130321230