Add option --action_env and make BuildConfiguration declare environment dependencies
This option will allow to specify which environment variables are to
be provided to the actions. Environment variables for the options will
become opt-in, i.e., only environment variables explicitly specified
will be provided to the actions. However, the full range of rc-files
will be able to nominate options to be added; to avoid dependency on
the invocation environment, absolute values can be provided as well.

As the configuration now no longer completely determines the action
environment extend it to also declare which environment variables are
to be taken from the client environment. This will be implemented in
follow-up patches.

The transition plan is that the newly added option takes precedence
over the environment variables added by the fragments. This is conservative,
as the new option is not yet used anywhere. Then the effect of the
fragments will be provided by rc-files, and finally, the setupShellEnvironment
Method will be removed from the fragments all together.

Also add some simple tests for static (i.e., independent of the client
environment) setting of action environment variables.

This is the first step towards implementing the design on [Specifying
environment variables for actions](http://bazel.io/designs/2016/06/21/environment.html).

--
Change-Id: I0ad36913b7d357787b4d69e341926806b3fc61bf
Reviewed-on: https://bazel-review.googlesource.com/#/c/4241
MOS_MIGRATED_REVID=131044391