Rollback of commit da3d46d5b85a52f213529f5e487d933577f0dc26.
*** Reason for rollback ***

Breaks tachyon because of the fact that it has duplicate symbol errors but they get discarded when compiled into the same archive, which is not the case anymore as we are being more clever with dependencies....

*** Original change description ***

Instead of using the smallest group as the inputs for the generation actions, use instead the intersection of all the groups in which each proto appears. This further minimizes the average number of inputs per generation/compilation action.

--
MOS_MIGRATED_REVID=131099657