Merge xmlns decls in old manifest merger to help transition
The old manifest merger doesn't merge xmlns decls.
That could be a problem if a library has xmlns:tools
but the binary doesn't (and we don't strip tools:).
Library manifests may end up with more xmlns:tools
annotations while transitioning from the old manifest
merger to the new. It would be a problem if the same
manifest is used by both the old and new merger.
An alternative may be to strip the tools annotations
after merging w/ this old merger... other options?

--
MOS_MIGRATED_REVID=131332171