Avoid building android R class ijar / header jar
It is already ignored and the real jar used as
the compileTime jar in AndroidCommon. It's only
used by the ide-resolve output group. The real
jar is pretty close to the header jar in terms
of content (field inits deleted).

Tradeoff: The real jar is ~33% bigger (w/ same
compression level), vs save build time and space
by skipping the hjar actions and artifacts.

At android_binary level, the hjar action can take
~40 seconds in some cases. At the library level,
the hjar time is pretty tiny.

hjars are not compressed. We could compress the
real jars to make up for the extra 30% size
(and beat the hjar size), but that adds ~25%
elapsed time to do the compression (e.g.,
10s vs 7.5s). Compression can shrink the jar
2-3x (37MB vs 100MB).

--
MOS_MIGRATED_REVID=131588283