Change default apple toolchain from darwin_x86_64 to ios_x86_64.
This means that in the absence of "--config=" (or the "--cpu=" and "--ios_cpu=" flags that it indicates), c++ code using //tools/osx/crosstool will default to being built for ios_x86_64 instead of darwin_x86_64.

This is necessary to ensure that once the crosstool is used for objc, the default behavior of objc_library building for ios_x86_64 will be maintained.

--
MOS_MIGRATED_REVID=132326009