Move the GetOutputRoot function to the WorkspaceLayout module.
The result value of GetOutputRoot does not depend on the startup
options: it only depends on the environment and/or the hardcoded values
for Bazel and Blaze.  Therefore, put it in the WorkspaceLayout module
just as we did for all other similar functions.

The fact that GetOutputRoot was part of BlazeStartupOptions was the root
cause behind the rollback of commit b517de5f0a727694d5d8f1be96f5d8c66c423a8b: in particular, that CL
silently added a virtual call to the GetOutputRoot method from the
constructor of the superclass, and this invokes undefined behavior
because the class has not yet been fully constructed.  This caused Blaze
to have incorrect values for the output_root.  By moving the function
out, we'll be able to roll that CL forward as it originally was.

As part of this change, add unit tests for the value of output_root
under various scenarios.  These would have caught the discrepancy
introduced by that CL.

--
MOS_MIGRATED_REVID=133056251